Move indexing override to overrides directory. Remove methods we don't override
